refactor(tests): consolidate duplicate test suites into nested directories (#254)
